comments methods finished, answer poll remaining on server

+++ Client/fair_politics/src/components/DiscussionCard.js
@@ -17,6 +17,10 @@ const DiscussionCard = ({ item, inProfile }) => {
   const [commentsButtonId, setCommentsButtonId] = useState(0);
   const [commentsButton, setCommentsButton] = useState(false);
   const [comment, setComment] = useState("");
+  const [commentEdit, setCommentEdit] = useState("");
+  const [editCommentForm, setEditCommentForm] = useState(false);
+  const [commentsEditButtonId, setCommentsEditButtonId] = useState(0);
+
   // const [height, setHeight] = useState(0);
   // const [onEdit, setOnEdit] = useState(false);
 
@@ -40,6 +44,7 @@ const DiscussionCard = ({ item, inProfile }) => {
       // setHeight(item.comments.length * 30);
     }
   };
+
   const LikeDiscussion = async () => {
     const details = {
       post_id: item.post_id,
@@ -61,10 +66,10 @@ const DiscussionCard = ({ item, inProfile }) => {
     const comment_details = {
       post_id: item.post_id,
       user_id: user_details.user_id,
-      comment: comment
+      comment: comment,
     };
     console.log(comment_details);
-    await fetch(`http://localhost:4000/api/add_like_to_discussion`, {
+    await fetch(`http://localhost:4000/api/add_comment`, {
       method: "POST",
       headers: { "Content-Type": "application/json" },
       body: JSON.stringify(comment_details),
@@ -72,25 +77,35 @@ const DiscussionCard = ({ item, inProfile }) => {
       .then((res) => res.json())
       .then((json) => console.log(json))
       .catch((err) => console.error(err));
-    setComment('');
+    setComment("");
   };
 
-  const deleteComment = async (e) => {
-    // const comment_details = {
-    //   post_id: item.post_id,
-    //   user_id: user_details.user_id,
-    //   comment: comment
-    // };
-    // console.log(comment_details);
-    await fetch(`http://localhost:4000/api/delete_comment/:comment_id`, {
-      method: "DELETE",
+  const editComment = async (e, comment_id) => {
+    const newComment = {
+      comment: commentEdit,
+    };
+    await fetch(`http://localhost:4000/api/update_comment/${comment_id}`, {
+      method: "PUT",
       headers: { "Content-Type": "application/json" },
-      body: JSON.stringify(),
+      body: JSON.stringify(newComment),
     })
       .then((res) => res.json())
       .then((json) => console.log(json))
       .catch((err) => console.error(err));
-    setComment('');
+    setComment("");
+    setEditCommentForm(false);
+  };
+
+  const deleteComment = async (e, comment_id) => {
+    if (window.confirm("Are you sure you want to delete this comment?")) {
+      await fetch(`http://localhost:4000/api/delete_comment/${comment_id}`, {
+        method: "DELETE",
+      })
+        .then((res) => res.json())
+        .then((json) => console.log(json))
+        .catch((err) => console.error(err));
+      setComment("");
+    }
   };
 
   const editDiscussion = (e) => {
@@ -100,7 +115,7 @@ const DiscussionCard = ({ item, inProfile }) => {
 
   const deleteDiscussion = async (e) => {
     setLoading(true);
-    if (window.confirm("Are you sure you want to delete this account?")) {
+    if (window.confirm("Are you sure you want to delete this post?")) {
       await fetch(
         `http://localhost:4000/api/delete_discussion/${item.post_id}`,
         {
@@ -115,7 +130,7 @@ const DiscussionCard = ({ item, inProfile }) => {
 
   return (
     <div style={styles.head}>
-      <Card style={{ height: 400, width: 600 }}>
+      <Card style={{ height: 1000, width: 600 }}>
         {inProfile && (
           <CardContent>
             <FaRegEdit onClick={(e) => editDiscussion(e)} />
@@ -148,11 +163,49 @@ const DiscussionCard = ({ item, inProfile }) => {
         </CardContent>
         {commentsButton && commentsButtonId === item.post_id && (
           <CardContent>
-            {item.comments.map((comment) => (
-              <div key={comment.comment_id} style={styles.comment}>
-                {comment.comment}
-              </div>
-            ))}
+            {item.comments.map((comment) => {
+              return (
+                <div key={comment.comment_id} style={styles.comment}>
+                  {comment.comment}
+                  {comment.user_id_comment === user_details.user_id && (
+                    <>
+                      <FaRegEdit
+                        onClick={(e) => {
+                          setEditCommentForm(!editCommentForm);
+                          setCommentsEditButtonId(comment.comment_id);
+                          setCommentEdit(comment.comment);
+                        }}
+                      />
+                      <FaTrashAlt
+                        onClick={(e) => deleteComment(e, comment.comment_id)}
+                      />
+                      {editCommentForm &&
+                        commentsEditButtonId === comment.comment_id && (
+                          <CardContent style={styles.formStyle}>
+                            <TextField
+                              // helperText='Tagged elected officials - Example: @israel israel @other person'
+                              id='standard-basic'
+                              // variant='standard'
+                              label='Update Comment'
+                              type='text'
+                              value={commentEdit}
+                              onChange={(e) => setCommentEdit(e.target.value)}
+                              // placeholder='add comment'
+                            />
+                            <Button
+                              variant='contained'
+                              onClick={(e) =>
+                                editComment(e, comment.comment_id)
+                              }>
+                              Submit
+                            </Button>
+                          </CardContent>
+                        )}
+                    </>
+                  )}
+                </div>
+              );
+            })}
           </CardContent>
         )}
         <CardContent style={styles.formStyle}>
@@ -224,40 +277,10 @@ const styles = {
     // right: 70,
   },
   comment: {
-    position: "relative",
-    top: 20,
-    // right:105,
+    // position: "relative",
+    // top: 20,
+    // // right:105,
   },
 };
-
-const DisCard = styled.div`
-  ${
-    "" /* background: linear-gradient(135deg, rgb(11,15,67) 5%,rgb(27,100,221) ); */
-  }
-  ${"" /* width: 14px; */}
-  ${"" /* height: 24px; */}
-  ${
-    "" /* background-image: url(data:image/svg+xml;utf8,%3Csvg%20viewBox%3D%220%200%2024%2024%22%20version%3D%221.1%22%20xmlns%3D%22http%3A//www.w3.org/2000/svg%22%20xmlns%3Axlink%3D%22http%3A//www.w3.org/1999/xlink%22%3E%0A%20%20%20%20%3Cdefs%3E%0A%20%20%20%20%20%20%20%20%3Cpath%20d%3D%22M5%2C11%20C5%2C7.691%207.691%2C5%2011%2C5%20C14.309%2C5%2017%2C7.691%2017%2C11%20C17%2C14.309%2014.309%2C17%2011%2C17%20C7.691%2C17%205%2C14.309%205%2C11%20M20.707%2C19.293%20L17.312%2C15.897%20C18.365%2C14.543%2019%2C12.846%2019%2C11%20C19%2C6.589%2015.411%2C3%2011%2C3%20C6.589%2C3%203%2C6.589%203%2C11%20C3%2C15.411%206.589%2C19%2011%2C19%20C12.846%2C19%2014.543%2C18.365%2015.897%2C17.312%20L19.293%2C20.707%20C19.488%2C20.902%2019.744%2C21%2020%2C21%20C20.256%2C21%2020.512%2C20.902%2020.707%2C20.707%20C21.098%2C20.316%2021.098%2C19.684%2020.707%2C19.293%22%20id%3D%22path-1%22%3E%3C/path%3E%0A%20%20%20%20%3C/defs%3E%0A%20%20%20%20%3Cg%20id%3D%22search%22%20stroke%3D%22none%22%20stroke-width%3D%221%22%20fill%3D%22none%22%20fill-rule%3D%22evenodd%22%3E%0A%20%20%20%20%20%20%20%20%3Cmask%20id%3D%22mask-2%22%20fill%3D%22white%22%3E%0A%20%20%20%20%20%20%20%20%20%20%20%20%3Cuse%20xlink%3Ahref%3D%22%23path-1%22%3E%3C/use%3E%0A%20%20%20%20%20%20%20%20%3C/mask%3E%0A%20%20%20%20%20%20%20%20%3Cuse%20id%3D%22%uD83C%uDFA8-Icon-%u0421olor%22%20fill%3D%22%230D1C2E%22%20fill-rule%3D%22nonzero%22%20xlink%3Ahref%3D%22%23path-1%22%3E%3C/use%3E%0A%20%20%20%20%3C/g%3E%0A%3C/svg%3E); */
-  }
- ${"" /* background-position: 50% 50%; */}
- ${"" /* background-repeat: no-repeat; */}
-  padding-left: 2rem;
-  background: transparent
-    linear-gradient(150deg, #025fdb 0%, #025fdb 35%, #0b3668 100%) 0% 0%
-    no-repeat padding-box;
-  box-shadow: 0 3px 20px rgb(0 0 0 / 8%);
-  height: 300px;
-  width: 500px;
-  margin: 10px;
-  ${"" /* position:absolute; */}
-  ${"" /* left:250px; */}
-  border-radius:30px;
-  ${"" /* display: flex; */}
-  ${"" /* flex-direction:row; */}
-  ${"" /* justify-content: space-between; */}
-  ${"" /* padding: 0.5rem calc((100vw - 1000px) / 2); */}
-  ${"" /* z-index: 10; */} /* Third Nav */
-  /* justify-content: flex-start; */
-`;
 
 export default DiscussionCard;
